PMM-4932 Split registry to handler
